+++ src/fortran_routing/mc_pylink_v00/MC_singleSeg_singleTS/pyMCsingleSegStime_NoLoop.f90
@@ -1,6 +1,6 @@
 module muskingcunge_interface
 
-use, intrinsic :: iso_c_binding, only: c_float
+use, intrinsic :: iso_c_binding, only: c_double
 use muskingcunge_module, only: muskingcungenwm
 
 implicit none
@@ -8,12 +8,12 @@ contains
 subroutine c_muskingcungenwm(dt, qup, quc, qdp, ql, dx, bw, tw, twcc,&
     n, ncc, cs, s0, velp, depthp, qdc, velc, depthc, ck, cn, X) bind(c)
 
-    real(c_float), intent(in) :: dt
-    real(c_float), intent(in) :: qup, quc, qdp, ql
-    real(c_float), intent(in) :: dx, bw, tw, twcc, n, ncc, cs, s0
-    real(c_float), intent(in) :: velp, depthp
-    real(c_float), intent(out) :: qdc, velc, depthc
-    real(c_float), intent(out) :: ck, cn, X
+    real(c_double), intent(in) :: dt
+    real(c_double), intent(in) :: qup, quc, qdp, ql
+    real(c_double), intent(in) :: dx, bw, tw, twcc, n, ncc, cs, s0
+    real(c_double), intent(in) :: velp, depthp
+    real(c_double), intent(out) :: qdc, velc, depthc
+    real(c_double), intent(out) :: ck, cn, X
 
     call muskingcungenwm(dt, qup, quc, qdp, ql, dx, bw, tw, twcc,&
     n, ncc, cs, s0, velp, depthp, qdc, velc, depthc, ck, cn, X)
